вынес кнопки в новые компоненты
